+++ scripts/translations/fetch.py
@@ -175,6 +175,12 @@ dir_keys = [
             ("google_code", ["config", "progress", "exchange"]),
         ],
     ),
+    (  # HOMEASSISTANT
+        f"{COMPONENTS}/homeassistant/translations/",
+        [
+            ("version", ["system_health", "info", "version"]),
+        ],
+    ),
     (  # UPDATE
         f"{COMPONENTS}/update/translations/",
         [
@@ -457,6 +463,9 @@ dir_keys = [
             ("description", ["ui.panel.config.automation.editor.description.label"]),
             ("manage_account", ["ui.panel.config.cloud.account.manage_account"]),
             ("configure", ["ui.panel.config.integrations.configure"]),
+            ("check_updates", ["ui.panel.config.updates.check_updates"]),
+            ("checking_updates", ["ui.panel.config.updates.checking_updates"]),
+            ("version_fmt", ["ui.panel.config.zwave_js.device_info.zwave_plus_version"]),
         ],
     ),
     (  # ONBOARDING